Update
Deel Sander implementeerd

+++ CubeTesten/JetPacL452REStm/Inc/Game.h
@@ -1,6 +1,7 @@
 #include "Highscore.h"
 #include "Sprite.h"
 #include "Object.h"
+#include "PData.h"
 
 #ifndef Game_H_
 #define Game_H_
@@ -19,7 +20,7 @@ public:
 	States GameState;
 
 	void removeObjectUni(uint8_t UniqueNr);
-	uint8_t addObject(Object * obj);
+	int addObject(Object * obj);
 	int ChangePosObjbyUniNr(uint8_t uninr, uint16_t x, uint16_t y);
 	bool changeSpriteNr(int,int); // Dangerous!!!!
 	//void showall();
